move validation and nullity handling to service layer

+++ src/main/java/pl/szotaa/snippr/snippet/service/SnippetService.java
@@ -4,21 +4,25 @@ import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.annotation.Validated;
 import pl.szotaa.snippr.snippet.domain.Snippet;
+import pl.szotaa.snippr.snippet.exception.SnippetNotFoundException;
 import pl.szotaa.snippr.snippet.repostiory.SnippetRepository;
 import pl.szotaa.snippr.user.domain.ApplicationUser;
 import pl.szotaa.snippr.user.service.ApplicationUserService;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @Service
+@Validated
 @AllArgsConstructor(onConstructor = @__({@Autowired}))
 public class SnippetService {
 
     private final SnippetRepository snippetRepository;
     private final ApplicationUserService applicationUserService;
 
-    public void save(Snippet snippet){
+    public void save(@Valid Snippet snippet){
 
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
         if(username != null){
@@ -29,19 +33,29 @@ public class SnippetService {
         snippetRepository.save(snippet);
     }
 
-    public Snippet getById(Long id){
-        return snippetRepository.findOne(id);
+    public Snippet getById(Long id) throws Exception {
+        Snippet found = snippetRepository.findOne(id);
+        if(found == null){
+            throw new SnippetNotFoundException(id);
+        }
+        return found;
     }
 
     public List<Snippet> getAll(){
         return null;
     }
 
-    public void update(Snippet snippet){
+    public void update(@Valid Snippet snippet) throws Exception {
+        if(!snippetRepository.exists(snippet.getId())){
+            throw new SnippetNotFoundException(snippet.getId());
+        }
         snippetRepository.save(snippet);
     }
 
-    public void delete(Long id){
+    public void delete(Long id) throws Exception {
+        if(!snippetRepository.exists(id)){
+            throw new SnippetNotFoundException(id);
+        }
         snippetRepository.delete(id);
     }
